Renamed glsl_util to util_3d. Moved functionality from Camera3d in opengl to a shared CameraBase class. Moved files that will be shared between vulkan and opengl to util_3d.
